Switching to default backgrounds instead of directb2s resources

+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/utils/FileUtils.java
@@ -1,5 +1,6 @@
 package de.mephisto.vpin.commons.utils;
 
+import edu.umd.cs.findbugs.annotations.Nullable;
 import org.apache.commons.io.FilenameUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,8 +14,8 @@ import java.text.DecimalFormat;
 public class FileUtils {
   private final static Logger LOG = LoggerFactory.getLogger(FileUtils.class);
 
-  public static boolean delete(File file) {
-    if(file.exists()) {
+  public static boolean delete(@Nullable File file) {
+    if(file != null && file.exists()) {
       if(file.delete()) {
         LOG.info("Deleted " + file.getAbsolutePath());
       }
